+++ src/user_registration/models/__init__.py
@@ -1,0 +1,8 @@
+from user_registration.models.registration import RegistrationRequest
+from user_registration.models.user import User
+
+
+__all__ = [
+    "RegistrationRequest",
+    "user"
+]